Fix bad clipping of error in DQN and DDPG.
The problem here was that we clipped the error in such a way that the
gradient would be zero everywhere outside of the clip region. This is
obviously not desired as pointed out by @karpathy and was thus fixed by
using the Huber loss.

For details, see
https://medium.com/@karpathy/yes-you-should-understand-backprop-e2f06eab496b
and https://github.com/devsisters/DQN-tensorflow/issues/16